# Conflicts:
#	FWLog.Data/Repository/GeneralCtx/NotaFiscalRepository.cs

+++ FWLog.Data/Models/LoteStatus.cs
@@ -5,7 +5,7 @@ namespace FWLog.Data.Models
     public class LoteStatus
     {
         [Key]
-        public int IdLoteStatus { get; set; }
+        public long IdLoteStatus { get; set; }
         public string Descricao { get; set; }
     }
 }
